Use "nestedtext" as syntax group prefix rather than "nt"
This is a nit, and I almost didn't bother, but I figured if it was going
to happen at all, it should come first.  The advantages of the full
prefix is that it prevents a hypothetical collision with another syntax
file, and it enables certain programmatic use cases, such as how
syntaxcomplete#Complete filters syntax groups by name when looking for
keywords to complete.  (This specific example is completely irrelevant
to NestedText as it doesn't have any keywords.)  It also makes it a bit
easier for users to remember where overrides they add to their color
scheme came from.

The principle disadvantage is longer line length, but the highlighting
here is simple enough that it's mostly a non-issue.