Only highlight tabs as errors
This is mostly a taste thing.  I personally find it distracting when
error highlighting intermittently pops in and vanishes while I'm typing.
The *lack* of highlighting is generally sufficient to convey that the
line is incomplete.  Reserve error highlighting for the unambiguous
mistakes.

A massive secondary benefit is that we can clean up a lot of the
eager whitespace highlighting, which is both simpler and will allow
background colors and text decorations like underline to look right in
color schemes that use them.